+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/android/orientations/AndroidOrientationsListCommand.kt
@@ -19,7 +19,7 @@ import java.nio.file.Paths
 )
 class AndroidOrientationsListCommand : Runnable {
     override fun run() {
-        println(AndroidCatalog.supportedOrientationsAsTable(AndroidArgs.load(Paths.get(configPath)).project))
+        println(AndroidCatalog.supportedOrientationsAsTable(AndroidArgs.loadOrDefault(Paths.get(configPath)).project))
     }
 
     @CommandLine.Option(names = ["-c", "--config"], description = ["YAML config file path"])